chore: modify utils
