+++ app/api/profile/web2/[handle]/route.ts
@@ -1,11 +1,11 @@
 import type { NextRequest } from "next/server";
 import { resolveIdentity } from "web3bio-profile-kit/utils";
 import { resolveUniversalHandle } from "../../[handle]/utils";
-import { getUserHeaders } from "@/utils/utils";
 import {
+  getUserHeaders,
   invalidIdentityResponse,
   parseResolvedIdentityHandle,
-} from "@/app/api/_shared/identity-route";
+} from "@/utils/utils";
 
 export async function GET(
   req: NextRequest,
